ref(metrics): move metrics api tag details endpoint to separate file
